functionally integrated version of feature extractor for local context
Core Concept class replaced with SHARPn-style IndentifiedAnnotation.
Sentence and Token types replaced with equivalent generic class.
Reader works with negation so far. We need to add flexibility to allow
for a single flag with multiple possible context states (that is,
mutually exclusive negation vs. historical vs. uncertain etc.) or
multiple labels on the same concept (that is, negation +/- flag
and not patient +/- flag and etc.). Parent type converted from
CAS Consumer to analysis engine so that is can be part of a
real pipeline as feature extractor or used as a utility (here)
to generate feature for training models.