noop support for brat ann and i2b2 2010 assertion (ast) files
We read in brat and ast files by checking for a parallel
text and .ann/.ast structured annotation file before adding
it to the queue of documents to process. Currently, the structured
info isn't actually read in yet so these two types are functionally
equivalent to plain text files alone.

Patch to writer to strip of the .xmi file endings on filenames
before generating the output filename (so we don't end up
with .xmi.xmi files).