fixed uncertainty default to be 0; added all attributes to XMI and ast
The default value for uncertainty should be 0 (for no uncertainty).
Then when uncertainty is mentioned, it is set to 1.

The ast file type was missing default values for generic.

The current cTAKES type system include historyOf, presumably for
historical annotations.  Added this type to XMI and ast files.

All attributes parsed for XMI files now.