bug fixes: sentence/token map mismatch; quotes in brat ann files
Two minor bug fixes for odd edge cases. For brat ann files, we
want to ignore all quotes as special characters. They don't enclose
special text fields for the faux csv reading we're doing.

In the AlAttrFeatureGen, there is an odd edge case when the sentence
map offsets don't match up with token map offsets. Likely, this is
because the sentence end officially lands outside of a token. We need to patch
this more explicitly with a robust log message. For now, we just make
sure to pull a valid entry.