updated pipeline.properties; uses liblinear to predict trained models
I updated the pipeline.properties.TEMPLATE file to show the new
key/value pairs being used by the system.

When a model file (and a few other properties) are set correctly,
the FeatureGen will also try to do a direct prediction based on
the features provided. Simple evaluation is printed at the end
of a corpus run. The predicted category/ies are written as
NOTE_NLP table concepts so that we can clearly distinguish
the internal/pre-processed annotation properties from the
model outputs.