FastText feature file generated; FastContext bug fixes
The fastText training file (one annotation per line with
label in the first column and the concept highlighted) is now
generated alongside the SVM/deep learning formatted file.

The feature directory is automatically created now (if not
present).

The setContext function was refactored (and its argument generation)
a bit to deal with a bug in subiterator for uimaFIT.  It's likely
slower than the previous algorithm but that can be improved upon.

Most importantly. the FastContext output is added at a feature
to the SVM and deep learning feature file.

the runAttr.sh was updated to include different label files depending
on whether we're training/predicting (using the combined labels)
or evaluating (against just the i2b2 labels).